Use 64 bit primitives to store values

+++ src/lang/ast.rs
@@ -17,7 +17,7 @@ pub enum Expr {
     Div(Box<Expr>, Box<Expr>),
 
     Var(String),
-    Num(f32),
+    Num(f64),
     Call(String, Vec<Expr>),
 }
